main2.py
'''

ANY CHANGES - BEFORE PRODCUTION VALUE (OF MAIN.PY), DO IT HERE
PASTE TO MAIN.PY LATER

'''

import asyncio
import os
import websockets
import json
import base64
from dotenv import load_dotenv
import argparse

from agent_utils2 import FUNCTION_MAP, synth_text_to_pcm16_wav, pcm16_wav_to_mulaw_8k_frames

load_dotenv()

def sts_connect():
    api_key = os.getenv("DEEPGRAM_API_KEY")
    if not api_key:
        raise Exception("DEEPGRAM_API_KEY not found.")
    
    sts_ws = websockets.connect(
        "wss://agent.deepgram.com/v1/agent/converse",  # can check another connection - docs
        subprotocols=["token", api_key]
    )
    
    return sts_ws


def load_config():
    with open("config.json", 'r') as f:
        return json.load(f)
    
    
async def handle_barge_in(decoded, twilio_ws, streamsid):
    if decoded['type'] == "UserStartedSpeaking":
        clear_message = {
            "event": "clear",
            "streamSid": streamsid
        }
        
        await twilio_ws.send(json.dumps(clear_message))


async def handle_text_message(decoded, twilio_ws, sts_ws, streamsid):
    await handle_barge_in(decoded, twilio_ws, streamsid)
    
    
    # STILL TO-DO: Add function calling


async def sts_sender(sts_ws, audio_queue):
    print("sts_sender started.")
    while True:
        chunk = await audio_queue.get()
        await sts_ws.send(chunk)


async def sts_receiver(sts_ws, twilio_ws, streamsid_queue):
    print("sts_receiver started.")
    streamsid = await streamsid_queue.get()
    
    async for message in sts_ws:
        if type(message) is str:
            print(message)
            decoded = json.loads(message)
            await handle_text_message(decoded, twilio_ws, sts_ws, streamsid)
            continue
        
        raw = message
        
        media_message = {
            "event": "media",
            "streamSid": streamsid,
            "media": {"payload": base64.b64decode(raw).decode("ascii")}
        }
        
        await twilio_ws.send(json.dumps(media_message))


async def enqueue_tts_text(text: str, audio_queue: asyncio.Queue, tts_model: str = None):
    """
    Synthesize 'text' to speech, convert to 8k μ-law, enqueue 20 ms frames
    so sts_sender() will stream them to Deepgram ASR.
    """
    cfg = load_config()  # reuse your config.json
    if tts_model is None:
        try:
            tts_model = cfg["agent"]["speak"]["provider"]["model"]
        except Exception:
            tts_model = "aura-2-thalia-en"

    wav = synth_text_to_pcm16_wav(text, model_from_config=tts_model)
    for frame in pcm16_wav_to_mulaw_8k_frames(wav):
        await audio_queue.put(frame)


async def twilio_receiver(twilio_ws, audio_queue, streamsid_queue):
    BUFFER_SIZE = 20 * 160
    inbuffer = bytearray(b"")
    
    async for message in twilio_ws:
        try:
            data = json.loads(message)
            event = data['event']
            
            if event == "start":
                print("Getting our StreamsIDs")
                start = data['start']
                streamsid = start['streamSid']
                streamsid_queue.put_nowait(streamsid)
                
            elif event == "connected":
                continue
            
            elif event == "media":
                media = data["media"]
                chunk = base64.b64decode(media["payload"])
                if media["track"] == "inbound":
                    inbuffer.extend(chunk)
                    
            elif event == "stop":
                break
            
            while len(inbuffer) >= BUFFER_SIZE:
                chunk = inbuffer[:BUFFER_SIZE]
                audio_queue.put_nowait(chunk)
                inbuffer = inbuffer[BUFFER_SIZE:]
                
        except:
            break


async def twilio_handler(twilio_ws):
    audio_queue = asyncio.Queue()
    streamsid_queue = asyncio.Queue()
    
    async with sts_connect() as sts_ws:
        config_message = load_config()
        await sts_ws.send(json.dumps(config_message))
        
        await asyncio.wait(
            [
                asyncio.ensure_future(sts_sender(sts_ws, audio_queue)),
                asyncio.ensure_future(sts_receiver(sts_ws, twilio_ws, streamsid_queue)),
                asyncio.ensure_future(twilio_receiver(twilio_ws, audio_queue, streamsid_queue))
            ]
        )
        
        await twilio_ws.close()
    


async def main():
    await websockets.serve(twilio_handler, "localhost", 5000)
    print("Started server")
    await asyncio.Future()
    
    
if __name__ == "__main__":
    asyncio.run(main())