pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "ipykernel>=7.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "websockets>=15.0.1",
]
